backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiosqlite
google-genai
python-dotenv
pydantic-settings
httpx
websockets
pyaudio
